Unify the styles / usage and theme of the select fields
